+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## 2025-07-02
+
+### Fixed
+- Pos now reset when starting a new file
+
+### Added
+- Image enhancement using PIL contrast & brightness settings for better
+- Full screen images without black bars
+- Startup messages now appear on RHS of screen, image on LHS
+- QR Code on RHS of startup screen for config URL
+
+## 2025-07-02
+
 ### Fixed
 - Skipping blank frames did not advance. Changed logic of main loop in update_display and use new function to get ffmpeg to find the next non-black frame.